Restructure dependency resolver test suite (#1395)
Without this change, IntelliJ 2021.3.2 does not detect individual test
cases, which makes them annoying to run individually.

+++ joern-cli/frontends/x2cpg/src/test/scala/io/joern/x2cpg/utils/dependency/DependencyResolverTests.scala
@@ -9,36 +9,44 @@ import java.nio.file.{Files, Path, Paths, StandardCopyOption}
 import java.util.Comparator
 
 class DependencyResolverTests extends AnyWordSpec with Matchers {
-  private class FixtureWithCopyDir(srcDir: Path) {
+  private class FixtureWithCopyDir(srcDir: Path, runningOnWindowsGitHubAction: Boolean = false) {
     def test(testFunc: collection.Seq[String] => Unit): Unit = {
-      val tmpDir = Files.createTempDirectory("DependencyResolverTests")
-      try {
-        Files.copy(srcDir, tmpDir, StandardCopyOption.REPLACE_EXISTING)
-        val dependenciesFiles = DependencyResolver.getDependencies(tmpDir)
-        testFunc(dependenciesFiles)
-      } finally {
-        Files
-          .walk(tmpDir)
-          .sorted(Comparator.reverseOrder[Path]())
-          .forEach(Files.delete(_))
+      if (runningOnWindowsGitHubAction) {
+        info("tests were cancelled because github actions windows doesn't support them for some unknown reason...")
+      } else {
+        val tmpDir = Files.createTempDirectory("DependencyResolverTests")
+        try {
+          Files.copy(srcDir, tmpDir, StandardCopyOption.REPLACE_EXISTING)
+          val dependenciesFiles = DependencyResolver.getDependencies(tmpDir)
+          testFunc(dependenciesFiles)
+        } finally {
+          Files
+            .walk(tmpDir)
+            .sorted(Comparator.reverseOrder[Path]())
+            .forEach(Files.delete(_))
+        }
       }
     }
   }
 
-  private class Fixture(content: String, fileName: String) {
+  private class Fixture(content: String, fileName: String, runningOnWindowsGitHubAction: Boolean = false) {
     def test(testFunc: collection.Seq[String] => Unit): Unit = {
-      val tmpDir = Files.createTempDirectory("DependencyResolverTests")
-      try {
-        val file = tmpDir.resolve(fileName)
-        Files.write(file, content.getBytes)
+      if (runningOnWindowsGitHubAction) {
+        info("tests were cancelled because github actions windows doesn't support them for some unknown reason...")
+      } else {
+        val tmpDir = Files.createTempDirectory("DependencyResolverTests")
+        try {
+          val file = tmpDir.resolve(fileName)
+          Files.write(file, content.getBytes)
 
-        val dependenciesFiles = DependencyResolver.getDependencies(tmpDir)
-        testFunc(dependenciesFiles)
-      } finally {
-        Files
-          .walk(tmpDir)
-          .sorted(Comparator.reverseOrder[Path]())
-          .forEach(Files.delete(_))
+          val dependenciesFiles = DependencyResolver.getDependencies(tmpDir)
+          testFunc(dependenciesFiles)
+        } finally {
+          Files
+            .walk(tmpDir)
+            .sorted(Comparator.reverseOrder[Path]())
+            .forEach(Files.delete(_))
+        }
       }
     }
   }
@@ -48,69 +56,69 @@ class DependencyResolverTests extends AnyWordSpec with Matchers {
     * few days, a proper debugging session will have to wait.
     */
   // TODO: remove this workaround
-  val isGithubActions = scala.util.Properties.envOrElse("GITHUB_ACTIONS", "false").toLowerCase == "true"
-  val isWindows       = scala.util.Properties.isWin
+  val isGithubActions                = scala.util.Properties.envOrElse("GITHUB_ACTIONS", "false").toLowerCase == "true"
+  val isWindows                      = scala.util.Properties.isWin
+  val isRunningOnWindowsGithubAction = isGithubActions && isWindows
 
-  if (isGithubActions && isWindows) {
-    info("tests were cancelled because github actions windows doesn't support them for some unknown reason...")
-  } else {
-    "test gradle dependency resolution for a simple `build.gradle`" in {
-      val fixture = new Fixture(
-        """
-          |apply plugin: 'java'
-          |repositories { mavenCentral() }
-          |dependencies { implementation 'log4j:log4j:1.2.17' }
-          |""".stripMargin,
-        "build.gradle"
-      )
+  "test gradle dependency resolution for a simple `build.gradle`" in {
+    val fixture = new Fixture(
+      """
+        |apply plugin: 'java'
+        |repositories { mavenCentral() }
+        |dependencies { implementation 'log4j:log4j:1.2.17' }
+        |""".stripMargin,
+      "build.gradle",
+      isRunningOnWindowsGithubAction
+    )
 
-      fixture.test { dependenciesFiles =>
-        dependenciesFiles.find(_.endsWith("log4j-1.2.17.jar")) should not be empty
-      }
+    fixture.test { dependenciesFiles =>
+      dependenciesFiles.find(_.endsWith("log4j-1.2.17.jar")) should not be empty
     }
+  }
 
-    "test gradle dependency resolution for a simple `build.gradle.kts`" in {
-      val fixture = new Fixture(
-        """
-          |plugins { kotlin("jvm") version "1.6.10" }
-          |repositories { mavenCentral() }
-          |dependencies { implementation("log4j:log4j:1.2.17") }
-          |""".stripMargin,
-        "build.gradle.kts"
-      )
+  "test gradle dependency resolution for a simple `build.gradle.kts`" in {
+    val fixture = new Fixture(
+      """
+        |plugins { kotlin("jvm") version "1.6.10" }
+        |repositories { mavenCentral() }
+        |dependencies { implementation("log4j:log4j:1.2.17") }
+        |""".stripMargin,
+      "build.gradle.kts",
+      isRunningOnWindowsGithubAction
+    )
 
-      fixture.test { dependenciesFiles =>
-        dependenciesFiles.find(_.endsWith("log4j-1.2.17.jar")) should not be empty
-      }
+    fixture.test { dependenciesFiles =>
+      dependenciesFiles.find(_.endsWith("log4j-1.2.17.jar")) should not be empty
     }
+  }
 
-    "test gradle dependency resolution for `build.gradle` using `kotlin-gradle-plugin`" in {
-      val fixture = new Fixture(
-        """
-          |buildscript {
-          |    repositories { mavenCentral() }
-          |    dependencies { classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.0" }
-          |}
-          |repositories { mavenCentral() }
-          |apply plugin: 'kotlin'
-          |dependencies { implementation 'log4j:log4j:1.2.17' }
-          |""".stripMargin,
-        "build.gradle"
-      )
+  "test gradle dependency resolution for `build.gradle` using `kotlin-gradle-plugin`" in {
+    val fixture = new Fixture(
+      """
+        |buildscript {
+        |    repositories { mavenCentral() }
+        |    dependencies { classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.0" }
+        |}
+        |repositories { mavenCentral() }
+        |apply plugin: 'kotlin'
+        |dependencies { implementation 'log4j:log4j:1.2.17' }
+        |""".stripMargin,
+      "build.gradle",
+      isRunningOnWindowsGithubAction
+    )
 
-      fixture.test { dependenciesFiles =>
-        dependenciesFiles.find(_.endsWith("log4j-1.2.17.jar")) should not be empty
-      }
+    fixture.test { dependenciesFiles =>
+      dependenciesFiles.find(_.endsWith("log4j-1.2.17.jar")) should not be empty
     }
+  }
 
-    "test gradle dependency resolution for simple Android app" in {
-      val androidAppDir = ProjectRoot.relativise("joern-cli/src/test/resources/testcode/SlimAndroid")
-      val fixture       = new FixtureWithCopyDir(Paths.get(androidAppDir))
-      fixture.test { dependenciesFiles =>
-        dependenciesFiles.filter(_.endsWith(".jar")) should not be Set()
-      }
-      // TODO: add test for `.aar` as soon as it's decided what to do about them
+  "test gradle dependency resolution for simple Android app" in {
+    val androidAppDir = ProjectRoot.relativise("joern-cli/src/test/resources/testcode/SlimAndroid")
+    val fixture       = new FixtureWithCopyDir(Paths.get(androidAppDir), isRunningOnWindowsGithubAction)
+    fixture.test { dependenciesFiles =>
+      dependenciesFiles.filter(_.endsWith(".jar")) should not be Set()
     }
+    // TODO: add test for `.aar` as soon as it's decided what to do about them
   }
 
   "test maven dependency resolution" in {
